Minor version select bugfixes
Whenever opening a new game, it will open in a new tab rather than in the same one.
Also fixed a bug in the mod selection. Basically if you selected a mod then left and went to another version that doesn't have that mod, it will not let you go to that version until you manually went back and deselected it. Now it automatically deselects all mods whenever you go back, as it's supposed to.